deepseekR1/requirements.txt
# Project dependencies
requests
pandas
python-dotenv
bibtexparser
rispy # For reading RIS files
# PyPDF2 # Optional: uncomment if PDF parsing is needed 